Added infinite scrolling to search page
