+++ src/app/library/LibraryMediaContentList.tsx
@@ -1,93 +1,23 @@
 'use client';
 
-import React, { useState } from 'react';
-import ContentCardView from '@components/layouts/library/content-card';
+import React from 'react';
 import { T_RawMediaContentFields } from 'types/media-content';
-import { determineFileType } from 'utils/helpers';
-import InfiniteScroll from 'react-infinite-scroll-component';
-import { MEDIA_CONTENT_LIMIT } from '../../components/library/constants';
-import { fetchMedia } from '../../components/library/fetchMedia';
+import { fetchMedia } from 'fetchers/library/fetchMedia';
 import useSearchFilters from '@hooks/useLibrary/useSearchFilters';
-import { Spinner } from 'flowbite-react';
+import { useLibraryInfiniteScroll } from '../../hooks/useLibrary/useLibraryInfiniteScroll';
+import { LibraryInfiniteScrollList } from '../../components/library/LibraryInfiniteScrollList';
 
 export default function LibraryMediaContentList({
   initialMediaContent,
 }: {
   initialMediaContent: T_RawMediaContentFields[];
 }) {
-  const [mediaContent, setMediaContent] = useState(initialMediaContent);
-  const [offset, setOffset] = useState(MEDIA_CONTENT_LIMIT);
-  let [hasMore, setHasMore] = useState(true);
-  let [error, setError] = useState(false);
-
   let filters = useSearchFilters();
 
-  useEffect(() => {
-    setMediaContent(initialMediaContent);
-    setOffset(MEDIA_CONTENT_LIMIT);
-    setHasMore(true);
-    setError(false);
-  }, [initialMediaContent]);
-
-  const loadMore = async () => {
-    try {
-      let result = await fetchMedia(offset, filters);
-
-      if (result instanceof Error || !result.mediaContent || !(result.mediaContent instanceof Array)) {
-        setError(true);
-        setHasMore(false);
-      } else if (result.mediaContent.length === 0) {
-        setHasMore(false);
-      } else {
-        const newItems = (result as { mediaContent: T_RawMediaContentFields[] }).mediaContent;
-        setMediaContent((existing) => [...existing, ...newItems]);
-      }
-    } catch {
-      setError(true);
-      setHasMore(false);
-    }
-
-    setOffset((offset) => offset + MEDIA_CONTENT_LIMIT);
-  };
-
-  return (
-    <>
-      <InfiniteScroll
-        dataLength={mediaContent.length}
-        next={loadMore}
-        hasMore={hasMore}
-        loader={
-          <div className="flex flex-row justify-center items-center my-4">
-            <Spinner size="md" />
-          </div>
-        }
-        endMessage={
-          error && (
-            <p className="text-center my-4">
-              <b className="text-red-500">Failed to load additional elements</b>
-            </p>
-          )
-        }
-        scrollableTarget="scrollableDiv"
-      >
-        <div className="grid pb-2 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 2xl:grid-cols-5">
-          {mediaContent.map((item) => (
-            <div style={{ padding: '8px 0' }} key={item._id} className="gutter-row px-2 h-full">
-              <ContentCardView
-                url={`/library/media/${item._id}`}
-                image={
-                  item.thumbnailUrl ||
-                  (item.file_url && determineFileType(item.file_url) === 'image'
-                    ? item.file_url
-                    : '/assets/images/no picture yet.svg')
-                }
-                title={item.name}
-                description={item.description}
-              />
-            </div>
-          ))}
-        </div>
-      </InfiniteScroll>
-    </>
+  let { error, hasMore, loadMore, mediaContent } = useLibraryInfiniteScroll(
+    initialMediaContent,
+    async (offset) => await fetchMedia(offset, filters),
   );
+
+  return <LibraryInfiniteScrollList mediaContent={mediaContent} loadMore={loadMore} hasMore={hasMore} error={error} />;
 }
